% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoPhase.R
\docType{package}
\name{a-OncoPhase}
\alias{OncoPhase-package}
\alias{a-OncoPhase}
\alias{a-OncoPhase-package}
\title{OncoPhase: An R package for somatic mutations cellular prevalence quantification using haplotype phasing.}
\description{
OncoPhase uses haplotype phase information to accurately compute mutational cellular prevalence. OncoPhase utilizes three sources of information: the phasing information, the copy number variation, and the allele counts.  It takes as input a combination of phased SNV and SNP allele-specific sequence read counts and local allele-specific copy numbers to determine the proportion of cells harboring the SNV and compute specific and detailed mutation cellular prevalence for each of the following groups of cells:
}
\details{
\describe{
      \item{Germ}{ Germline cells having a normal genotype with no mutations and no copy number alteration at the considered locus. }
      \item{Alt}{ Cells harboring one  alternative between the two somatic alterations. That is either only the SNV if C=1 (SNV occurred before SCNA) or only the SCNA if C=0 (SNV occurred after the SCNA).}
      \item{Both}{ Cells harboring both somatic alterations. That is the SNV and the SCNA}
   }

OncoPhase can compute the mutation cellular prevalence under three different modes : PhasedSNP, FlankingSNP and SNVOnly.



The main functions for somatic mutation cellular prevalence computation are    \code{\link{getPrevalence}} ,   \code{\link{getSamplePrevalence}}  and  \code{\link{getMultiSamplesPrevalence}} .   For more detailed information on usage, see the package vignette, by typing
\code{vignette("OncoPhase")}. All support questions should be emailed to the authors.
}
\author{
Donatien Chedom-Fotso, Ahmed Ahmed, Christopher Yau.
}
\references{
OncoPhase reference:

Chedom-Fotso Donatien, Ahmed Ashour Ahmed, and Christopher Yau. "OncoPhase: Quantification of somatic mutation cellular prevalence using phase information." bioRxiv (2016): 046631.
}
\keyword{package}

