% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoPhase.R
\docType{package}
\name{aOncoPhase}
\alias{OncoPhase-package}
\alias{aOncoPhase}
\alias{aOncoPhase-package}
\title{OncoPhase package for somatic mutations cellular prevalence quantification using haplotype phasing}
\description{
The main function for somatic mutation cellular prevalence computation are    \code{\link{getPrevalence}} and  \code{\link{getPrevalenceMultiSamples}} . \code{\link{getPrevalence}}  compute the cellular prevalence at a particular mutations under 4 distincts modes : PhasedSNP, FlankingSNP, OptimalSNP and SNVOnly.  \code{\link{getPrevalenceMultiSamples}}  computes the cellular prevalence of a list of mutations located at a given region of the genome. It can also work on a whole genome scale.
See the manual and examples at \code{\link{getPrevalence}} and  \code{\link{getPrevalenceMultiSamples}}  for more details.
}
\details{
The function \code{\link{getPrevalenceLinear}} compute the prevalence of a given mutation by directly solving the linear system associated to the model.

Input data for simple case studies can be generated with the function \code{\link{build_casestudy}}.

The package include experimental data for chromosome 22 from a  patients retrieved from a parallel clinical study.(see \code{\link{chr22_11152}} )

For more detailed information on usage, see the package vignette, by typing
\code{vignette("OncoPhase")}. All support questions should be emailed to the authors.
}
\author{
Donatien Chedom-Fotso, Ahmed Ahmed, Christopher Yau.
}
\references{
OncoPhase reference:

OncoPhase: A package for computing Somatic Mutation cellular Prevalence in cancer using haplotype phasing. Bioinformatics 2016. Submitted

OncoPhase reference:

” Ovarian cancer haplotype sequencing reveals ubiquitous SOX2 overexpression in the premalignant fallopian tube epithelium”
}
\keyword{package}

