% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoPhase.R
\name{build_casestudy}
\alias{build_casestudy}
\title{Build the input data matrix for a case study}
\usage{
build_casestudy(lambda_G, mu_G, lambda_S, mu_S, phi_G, major_cn, minor_cn)
}
\arguments{
\item{lambda_G}{: count of allele suporting the variant sequence of  the Germline SNP}

\item{mu_G}{: count of allele suporting the reference sequence  of  the Germline SNP}

\item{lambda_S}{: count of allele suporting the variant sequence of  the somatic mutation}

\item{mu_S}{: count of allele suporting the reference sequence  of  the somatic mutation}

\item{phi_G}{: Estimated fraction of cells affected by the CNV (1- normal genotype cell fraction)}

\item{minor_cn}{: Minor copy number at the locus of the mutation}

\item{major_cn:}{Major copy number at the locus of the mutation}

\item{snp_allelecount_df}{A data frame containing for each mutation the  allelic 
count of the variant at each tumour. Should contains at least Chrom (The mutation
 chromosome) , End (The mutation position) and IsGermline (is the mutation a Germline
  or Somatic)  among the first columnns.}

\item{ref_allelecount_df}{A data frame containing for each mutation the allelic count
of the reference at each tumour.Should contains at least Chrom (The mutation
 chromosome) , End (The mutation position) and IsGermline (is the mutation a Germline
  or Somatic)  among the first columnns.}

\item{phasing_association_df}{A data frame containing for each somatic mutation, 
a colon separated list of germline SNP phased to it.}

\item{major_copynumber_df}{A data frame containing for each mutation the major
chromosonal copy number at each tumour sample.}

\item{minor_copynumber_df}{A data frame containing for each mutation the minor
chromosonal copy number at each tumour sample.}

\item{normalfraction_df}{A data frame containing for each mutation the fraction of
normal cell contamination at each tumour/sample of the study.}
}
\value{
A list containing the following data frames:

A list containing the following dataframes :
 \describe{
       \item{snp_allelecount_df}{A data frame containing the  count of allele suporting the variant sequence at the somatic and germline  mutations. Chrom is set to chr3, Position of the germline and somatic mutations are respectively set to 100 and 1000}
       \item{ref_allelecount_df}{A data frame containing the  count of allele suporting the reference sequence at the somatic and germline  mutation. Chrom is set to chr3, Position of the germline and somatic mutations are respectively set to 100 and 1000}
       \item{phasing_association_df}{A data frame containing phasing association}
        \item{major_copynumber_df}{A data frame containing the major copy number at the mutation locus}
        \item{minor_copynumber_df}{A data frame containing the minor copy number at the mutation locus}
       \item{normalfraction_df}{A data frame containing the proportion of cells with a normal genotype}
     }
}
\description{
This is a generic function to automatically build the five input data frame (snp_allelecount_df, ref_allelecount_df, phasing_association_df, major_copynumber_df,minor_copynumber_df,normalfraction_df) for a case study with one somatic mutation, one germline mutation and one tumour sample.
}
\examples{

# We reproduce here the case study No 6 of the paper
 #Build the input data 
 cs = build_casestudy(lambda_G=16, mu_G=8,lambda_S=14,mu_S=10,phi_G=4/8,major_cn=3,minor_cn=1  )
 #Run the case
prevalence=getPrevalence(cs$snp_allelecount_df, cs$ref_allelecount_df, cs$phasing_association_df,
                         cs$major_copynumber_df,cs$minor_copynumber_df,cs$normalfraction_df)
#print the result
print(prevalence)
#Chrom  End IsGermline Tumour1
#somaticM  chr3 1000          0    0.75

}

