% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoPhase_methods.R
\name{getMatricesPhasedSNP}
\alias{getMatricesPhasedSNP}
\title{Generate the matrices C, W and M from a set of parameters.}
\usage{
getMatricesPhasedSNP(varcounts_snv, refcounts_snv, major_cn, minor_cn,
  varcounts_snp, refcounts_snp, context, LocusCoverage = FALSE,
  CopyNumberFraction = FALSE)
}
\arguments{
\item{varcounts_snv}{A count of
alleles supporting the variant sequence of  the somatic mutation}

\item{refcounts_snv}{A count of
alleles supporting the reference sequence  of  the somatic mutation}

\item{major_cn}{major copy number  at the locus of the mutation}

\item{minor_cn}{minor copy number   at the locus of the mutation}

\item{varcounts_snp}{A count of
alleles supporting the variant sequence of  the Germline SNP}

\item{refcounts_snp}{A count of
alleles  supporting the reference sequence  of  the Germline SNP}

\item{context}{represents either the situation of a mutation which occurred after the CNV ("C1") or the context of a mutation which occurred before the CNV ("C2"). If not provided, the right context will be estimated from the input}

\item{LocusCoverage}{when set to TRUE, the SNV locus coverage is estimated to the average coverage of the phased SNP and the variant allele fraction is the ratio of the variant allele count over the estimated locus coverage.}

\item{CopyNumberFraction}{when set to TRUE, if the mode is PhasedSNP then copy number fraction is considered as input to the linear model is replacement of the phased SNP allele fraction.}
}
\value{
the matrices W, C and M for the linear system of prevalence computation.
}
\description{
This is a generic function to generate the matrices of the linear system (see the paper) from the allele counts and the copy number information.
}
\examples{

Matrices = getMatricesPhasedSNP(3, 10,2,1,8,5,"C1")

print(Matrices)
#$context
#[1] "C1"
#
#$W
#          SNP       SNV
#SNP 0.6153846 0.0000000
#SNV 0.0000000 0.2307692
#
#$M
#    Germ Alt Both
#SNP    1   2    2
#SNV    0   0    1
#
#$C
#    Germ Alt Both
#SNP    2   3    3
#SNV    2   3    3

}

