% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoPhase_methods.R
\name{getMatricesSNVOnly}
\alias{getMatricesSNVOnly}
\title{Generate the matrices C, W and M from a set of parameters under the mode "SNVOnly.}
\usage{
getMatricesSNVOnly(varcounts_snv, refcounts_snv, major_cn, minor_cn, context,
  sigma = NULL)
}
\arguments{
\item{varcounts_snv}{A count of
alleles supporting the variant sequence of  the somatic mutation}

\item{refcounts_snv}{A count of
alleles supporting the reference sequence  of  the somatic mutation}

\item{major_cn}{major Copy number  at the locus of the mutation}

\item{minor_cn}{minor copy number   at the locus of the mutation}

\item{context}{represents either the situation of a mutation which occurred after the CNV ("C1") or the context of a mutation which occurred before the CNV ("C2"). If not provided, the right context will be estimated from the input}

\item{sigma}{Copy number of the parental chromosome harboring the mutation.}
}
\value{
the matrices W, C and M for the linear system of prevalence computation.
}
\description{
This is a generic function to generate the matrices of the linear system (see the paper) from the allele counts and the copy number information under the SNVOnly mode.
}
\examples{

Matrices = getMatricesSNVOnly(3,10,2,1,"C1")

print(Matrices)
#$context
#[1] "C1"
#
#$W
#SNV
#SNV 0.2307692
#
#$M
#Germ Alt Both
#SNV    0   0    1
#
#$C
#Germ Alt Both
#SNV    2   3    3

}

