% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoPhase_methods.R
\name{getMultiSamplesPrevalence}
\alias{getMultiSamplesPrevalence}
\title{Somatic mutations cellular prevalence computation using haplotype phasing on a multiple sample study.}
\usage{
getMultiSamplesPrevalence(snp_allelecount_df, ref_allelecount_df,
  major_copynumber_df, minor_copynumber_df, mode = "PhasedSNP",
  phasing_association_df = NULL, NormalCellContamination = NULL,
  nbFirstColumns = 3, tumoursamples = NULL, region = NULL,
  detail = TRUE, LocusRadius = 10000, LocusCoverage = TRUE,
  SomaticCountAdjust = TRUE)
}
\arguments{
\item{snp_allelecount_df}{A data frame containing for each mutation the  allelic 
Counts of the variant at each tumor samples. The data frame should contain at least the following three columns among its firsts columns: Chrom (The mutation
 chromosome) , Pos or End (The mutation position) and IsGermline (is the mutation a germline
  or somatic mutation).}

\item{ref_allelecount_df}{A data frame containing for each mutation the allelic count
 of the reference at each tumor sample. The data frame should contain at least the following three columns among its firsts columns:  Chrom (The mutation
chromosome) , Pos or End  and IsGermline (is the mutation a Germline
   or Somatic mutation)}

\item{major_copynumber_df}{A data frame containing for each mutation, its  major
chromosomal copy number at each tumor samples. Should contain at least the following three columns among its firsts columns:  Chrom (The mutation
chromosome) , Pos or End (The mutation position) and IsGermline (is the mutation a Germline
   or Somatic mutation)}

\item{minor_copynumber_df}{A data frame containing for each mutation the minor
 chromosomal copy number at each tumor samples. Should contain at least the following three columns among its firsts columns:  Chrom (The mutation
chromosome) , Pos or End (The mutation position) and IsGermline (is the mutation a Germline
   or Somatic mutation)}

\item{mode}{The mode under which the prevalence is computed  (default : PhasedSNP , alternatives modes  are FlankingSNP, OptimalSNP and SNVOnly).  Can also be provided as a numeric 0=SNVOnly, 1= PhasedSNP, 2=FlankingSNP and 3 = OptimalSNP}

\item{phasing_association_df}{A data frame containing for each somatic mutation, 
a colon separated list of germline SNP phased to it.}

\item{NormalCellContamination}{If provided, represents the rate of normal cells contaminations in the experiment.}

\item{nbFirstColumns}{Number of first columns in snp_allelecount_df to reproduce in
the output dataframe e.g: Chrom, Pos, Vartype. Columns from  nbFirstColumns +1 to the last column should contains the information needed for the prevalence computation at each tumor sample}

\item{tumoursamples}{The list of tumor samples to consider for the prevalence
computation.  These samples should be present as column headers in the data frame
  snp_allelecount_df, ref_allelecount_df, major_copynumber_df,minor_copynumber_df. If not provided, the headers from nbFirstColumns + 1 to 
  the last column of snp_allelecount_df are retrieved and their  intersection with the
other inputted data frames headers is considered}

\item{region}{The region of the genome to consider for the prevalence computation  in the format chrom:start-end 
e.g "chr22:179800-98767}

\item{detail}{when set to TRUE, a detailed output is generated containing, the context and the detailed prevalence for each group of cells (germline cells, cells affected by one of the two genomic alterations (SNV or CNV) but not both, cells affected by  both copy number alteration and SNV ).}

\item{LocusRadius}{Only phased SNPs located within LocusRadius bp from the somatic mutation will be considered.}

\item{LocusCoverage}{when set to TRUE, the SNV locus coverage is estimated to the average coverage of the phased SNP and the variant allele fraction is the ratio of the variant allele count over the estimated locus coverage.}

\item{SomaticCountAdjust}{when set to TRUE, varcounts_snv and refcounts_snv might be adjusted if necessary so that they meet the rules varcounts_snv <= varcounts_snp, refcounts_snv >= refcounts_snp and varcounts_snv + refcounts_snv ~ Poiss(varcounts_snp + refcounts_snp). Not used if mode=SNVOnly.}
}
\value{
A data frame containing :
 \describe{
       \item{}{Column 1 to NbFirstcolumn of the input data frame snp_allelecount_df. 
       This will generally include the chromosome and the position of the mutation plus
       any other columns to report in the prevalence data frame (e.g REF, ALL, ...) }
        \item{}{One column per tumor sample reporting the prevalence of the mutation 
        at each samples}
     }
}
\description{
This is a generic function to compute the cellular prevalence of somatic mutations in
 cancer using haplotype phasing.  The function applies the model to a range of mutations located at a given genomic region or at the whole genome scale. The model computes the prevalence of a somatic
  mutation relatively to close and eventually phased germline mutations. It uses three sources
   of information as input : The allelic counts, the phasing information and the 
   copy number alteration.  Multiple tumor samples can be provided for the prevalence computation.
}
\examples{

#Example 1: Loading a simple example data set with two somatic mutations, 5 germlines SNP
# and 3 tumor samples
data(simpleExample2)
se=simpleExample2
prevalence_df=getMultiSamplesPrevalence(se$snp_allelecount_df, se$ref_allelecount_df,
 se$major_copynumber_df,se$minor_copynumber_df,phasing_association_df=se$phasing_association_df, )
print(prevalence_df)

#Chrom     End IsGermline  Tumour1        Tumour2        Tumour3
#mutation2  chr2 3003000          0 C2:0|0|1 C2:0.15|0|0.85 C2:0.12|0|0.88
#mutation6  chr2 4008000          0 C1:1|0|0       C1:1|0|0 C2:0|0.24|0.76

#Example 2 : Computing somatic mutation cellular prevalence on chromosome 15 of  patient XYZ101 
# (data created)

data("chr22_XYZ101")
ds=chr22_XYZ101
masterprevalence_df=getMultiSamplesPrevalence(ds$snp_allelecount_df, ds$ref_allelecount_df,
  ds$major_copynumber_df,ds$minor_copynumber_df,phasing_association_df = ds$phasing_association_df,
   nbFirstColumns=6,detail=FALSE)
print(head(masterprevalence_df))

data("chr18_XYZ101")
df=chr18_XYZ101
masterprevalence_df=getMultiSamplesPrevalence(df$snp_allelecount_df, df$ref_allelecount_df,
 df$major_copynumber_df,df$minor_copynumber_df,phasing_association_df=df$phasing_association_df, 
 nbFirstColumns=6, region="chr18:10000000-80000000")
print(head(masterprevalence_df))



#'@seealso \\code{\\link{getPrevalence}}
}

