% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoPhase_methods.R
\name{getPhasedSNPPrevalenceGeneral}
\alias{getPhasedSNPPrevalenceGeneral}
\title{Compute detailed prevalence at a single mutation point under the Phased SNP mode using the General (exact) formula}
\usage{
getPhasedSNPPrevalenceGeneral(lambda_S, mu_S, major_cn, minor_cn, lambda_G,
  mu_G, cnv_fraction = NULL, min_cells = 1, min_alleles = 1,
  SameTumour = TRUE)
}
\arguments{
\item{lambda_S}{: A count (or a vector of counts) of 
alleles supporting the variant sequence of  the somatic mutation}

\item{mu_S}{: A count (or a vector of counts) of 
alleles supporting the reference sequence  of  the somatic mutation}

\item{major_cn}{major copy number (or a vector ) 
at the locus of the mutation}

\item{minor_cn}{: minor copy number (or a vector )
at the locus of the mutation}

\item{lambda_G}{A count (or a vector of counts if multiple samples) of 
alleles supporting the variant sequence of  the Germline SNP}

\item{mu_G}{A count (or a vector of counts) of 
alleles  supporting the reference sequence  of  the Germline SNP}

\item{cnv_fraction}{If provided, represents the fraction of cells affected by the copy number alteration. This value, if not provided, is computed from the allelic count information and copy number information. Default NULL}

\item{SameTumour}{In the case of a multiple sample computation, specify if the samples are originating from the same tumour.}

\item{formula}{Can be either "Matrix" either "General", specify if the prevalence should be computed using the linear form formula or the General form formula. Default "Matrix"}

\item{detail}{In case form="Matrix", when set to TRUE, a detailed output is generated containing, the context and the detailed prevalence for each group of cells (germline cells (Germ), cells affected by one of the two genomic alterations (Alt), cells affected by  by both genomic alterations (Both) ).}
}
\value{
\describe{
       \item{}{ if form="general", the function return a numerical value representing the prevalence at the somatic mutation.}
       \item{}{ if form="matrix", the function return a list containing the following data frames:
 \describe{
       \item{Context}{ The associated context}
       \item{Prevalence}{The computed prevalence}
       \item{DetailedPrevvalence}{Detailed prevalence for each of the three genotype groups separated by "|". The three groups are Germline mutations, mutations harboring one of the two alterations (CNV or SNP) mutations harboring both alterations }
       
     }
      }
      }
}
\description{
This is a generic function to compute the prevalence at a single somatic mutation point using a phased Germline SNP.
}
\examples{

# We reproduce here the case study No 6 of the paper
#prevalence=getPhasedSNPPrevalenceGeneral(lambda_S=14,mu_S=10,major_cn=3,minor_cn=1, lambda_G=16, mu_G=8, cnv_fraction=4/8 )
 
}
\seealso{
\code{\link{getPrevalence}}
}

