% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoPhase_methods.R
\name{getPrevalence}
\alias{getPrevalence}
\title{Computes cellular prevalence at a single mutation point}
\usage{
getPrevalence(varcounts_snv, refcounts_snv, major_cn, minor_cn,
  varcounts_snp = NULL, refcounts_snp = NULL, detail = 0,
  mode = "PhasedSNP", Trace = FALSE, LocusCoverage = TRUE,
  SomaticCountAdjust = TRUE, NormalCellContamination = NULL)
}
\arguments{
\item{varcounts_snv}{A count (or a vector of counts  if multiple samples ) of 
alleles supporting the variant sequence of  the somatic mutation}

\item{refcounts_snv}{A count (or a vector of counts  if multiple samples ) of 
alleles supporting the reference sequence  of  the somatic mutation}

\item{major_cn}{major copy number (or a vector  if multiple samples ) 
at the locus of the mutation}

\item{minor_cn}{minor copy number (or a vector  if multiple samples)
at the locus of the mutation}

\item{varcounts_snp}{A count (or a vector of counts  if multiple samples) of 
alleles supporting the variant sequence of  the Germline SNP}

\item{refcounts_snp}{A count (or a vector of counts  if multiple samples) of 
alleles  supporting the reference sequence  of  the Germline SNP}

\item{detail}{when set to 0, the function simply output the cellular prevalence of the somatic mutation. if set to 1,  a detailed output is generated containing:
 \describe{
       \item{Context}{ The associated context (C=1 or C=2) }
       \item{Prevalence}{The computed cellular prevalence}
       \item{DetailedPrevvalence}{the detailed prevalence for each group of cells (germline cells (Germ), cells affected by one of the two genomic alterations (Alt), cells affected  by both genomic alterations (Both)}
       \item{ResidualNorm}{Norm of the residuals when an approximation is performed}
       \item{CondensedPrevalence}{A colon separated list of the above fields (Context, Prevalence, Detailedprevalence and ResidualNorm). The detailed prevalence are separated by "|"}
}
  if detail is set to 2, the function outputs the CondensedPrevalence field above.}

\item{mode}{The mode under which the prevalence is computed  (default : PhasedSNP , alternatives methods  are FlankingSNP and SNVOnly). Can also be provided as a numeric 0=SNVOnly, 1= PhasedSNP, 2=FlankingSNP}

\item{Trace}{if set to TRUE, print the trace of the computation.}

\item{LocusCoverage}{when set to TRUE, the SNV locus coverage is estimated to the average coverage of the phased SNP and the variant allele fraction is the ratio of the variant allele count over the estimated locus coverage.}

\item{SomaticCountAdjust}{when set to 1, varcounts_snv and refcounts_snv might be adjusted if necessary so that they meet the rules varcounts_snv <= varcounts_snp, refcounts_snv >= refcounts_snp and varcounts_snv + refcounts_snv ~ Poiss(varcounts_snp + refcounts_snp). Not used if mode=SNVOnly,}

\item{NormalCellContamination}{If provided, represents the rate of normal cells contaminations in the experiment.}
}
\value{
The cellular prevalence if detail =0, a detailed output if detail = 1, and a condensed output if detail =2. See the usage of the parameter detail above.
}
\description{
This is a generic function to compute the cellular prevalence of a somatic mutation point using OncoPhase method. The method computes the prevalence of the somatic mutation relatively to phased nearby SNPs whose prevalence are known to be 1.  \code{\link{getPrevalence}} requires the allelic-information of the somatic mutation and the aggregated information of its Phased SNP but the function can also be run in the absence of phasing information (FlankingSNP mode) or nearby SNP (SNVOnly mode).
}
\details{
The method particularly exhibits an increase in the accuracy when the locus of the SNP is also affected by a somatic copy number alteration (SCNA). The method detect the temporal relationship between the two alterations (C1: SNV occurred after the SCNA; C2: SNV occurred before the SCNA) and computes the detailed prevalence of each of the following group of cells (if detail is set to TRUE) : 
      \describe{
       \item{Germ}{ Cells having a germline genotype  at the locus of the SNV. That is No SNV, no SCNA}
       \item{Alt}{ Cells having one alternative of  the two somatic alteration. That is either the SCNA, either the SNV not both.}
       \item{Both}{ Cells having both somatic alterations. That is the SNV and the SCNA}
    }

OncoPhase can be run under three modes:

\describe{
       \item{PhasedSNP}{ The prevalence is computed relatively to a Phased SNP}
       \item{FlankingSNP}{ In the absence of phasing information, the prevalence is computed relatively to a neighbor SNP located on the same locus with the somatic SNV. NA is returned if the prevalence cannot be resolved without knowing the phasing information between the SNP and the SNV.}
       \item{SNVOnly}{The prevalence is computed using only the SNV information without the usage of any nearby SNP}
    }
}
\seealso{
\code{\link{getPrevalence}},  \code{\link{getSamplePrevalence}},   \code{\link{getSinglePrevalence}}, \code{\link{getPrevalenceSNVOnly}}
}

