% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoPhase_methods.R
\name{getPrevalenceLinear}
\alias{getPrevalenceLinear}
\title{Compute the cellular prevalence of each group of cells}
\usage{
getPrevalenceLinear(lambda_S, mu_S, major_cn, minor_cn, lambda_G, mu_G, context,
  Trace = FALSE, LocusCoverage = 1)
}
\arguments{
\item{lambda_S}{: A count of 
alleles supporting the variant sequence of  the somatic mutation}

\item{mu_S}{: A count of 
alleles supporting the reference sequence  of  the somatic mutation}

\item{minor_cn}{: minor copy number (or a vector of copy number if multiple tumor samples)}

\item{lambda_G}{: A count of 
alleles supporting the variant sequence of  the Germline SNP}

\item{mu_G}{: A count of 
alleles  supporting the reference sequence  of  the Germline SNP}

\item{context}{represents either the situation of a mutation which occurred after the CNV ("C1") or the context of a mutation which occurred before the CNV ("C2"). If not provided, the right context will be estimated from the input}

\item{major_cn:}{major copy number at the locus of the mutation}
}
\value{
A list of the three cellular prevalence of each of the three groups of cells
}
\description{
This is a generic function to compute the detailed prevalence of a single mutation using the linear system of  the model.
}
\examples{

Prevalences = getPrevalenceLinear(8, 5,2,1,3,10,"C1")
 
  print(Prevalences)
# Germ  Alt Both 
# 0.4  0.0  0.6 

}
\seealso{
\code{\link{getPrevalence}},   \code{\link{getMatrices}}
}

