% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoPhase_methods.R
\name{getSamplePrevalence}
\alias{getSamplePrevalence}
\title{Somatic mutations cellular prevalence on a sample.}
\usage{
getSamplePrevalence(input_df, mode = "PhasedSNP", nbFirstColumns = 0,
  region = NULL, detail = TRUE, LocusCoverage = TRUE,
  SomaticCountAdjust = TRUE, NormalCellContamination = NULL)
}
\arguments{
\item{input_df}{A data frame containing for each mutations :
\describe{
       \item{varcounts_snv}{Alelle counts supporting the SNV}
       \item{refcounts_snv}{Alelle counts supporting the reference at the SNV locus}
       \item{major_cn}{Major copy number at the SNV locus}
       \item{minor_cn}{Minor copy number at the SNV locus}
       \item{varcounts_snp}{Alelle counts supporting the SNP}
       \item{refcounts_snp}{Alelle counts supporting the reference at the SNP}
     }}

\item{mode}{The mode under which the prevalence is computed  (Default : PhasedSNP , alternatives methods  are FlankingSNP, OptimalSNP,and SNVOnly).  Can also be provided as a numeric 0=SNVOnly, 1= PhasedSNP, 2=FlankingSNP and 3 = OptimalSNP}

\item{nbFirstColumns}{Number of first columns in input_df to reproduce in the output dataframe e.g: Chrom, Pos, Vartype. Columns from  nbFirstColumns +1 to the last column should contains the information needed for the prevalence computation.}

\item{region}{The region of the genome to consider for the prevalence computation  in the format chrom:start-end   e.g "chr22:179800-98767.}

\item{detail}{when set to TRUE, a detailed output is generated containing, the context and the detailed prevalence for each group of cells (germline cells, cells affected by one of the two genomic alterations (SNV or CNV) but not both, cells affected by  both copy number alteration and SNV ).}

\item{LocusCoverage}{when set to TRUE, the SNV locus coverage is estimated to the average coverage of the phased SNP and the variant allele fraction is the ratio of the variant allele count over the estimated locus coverage.}

\item{SomaticCountAdjust}{when set to TRUE, varcounts_snv and refcounts_snv might be adjusted if necessary so that they meet the rules varcounts_snv <= varcounts_snp, refcounts_snv >= refcounts_snp and varcounts_snv + refcounts_snv ~ Poiss(varcounts_snp + refcounts_snp). Not used if mode=SNVOnly,}

\item{NormalCellContamination}{If provided, represents the rate of normal cells contaminations in the experiment.}
}
\value{
A data frame containing :
 \describe{
       \item{}{Column 1 to NbFirstcolumn of the input data frame input_df. 
       This will generally include the chromosome and the position of the mutation plus
       any other columns to report in the prevalence dataframe (e.g REF, ALL, ...) }
        \item{}{and the following information}
  \describe{
       \item{Prev}{The Cellular Prevalence of the mutation}
       \item{Germ}{The proportion of cells with a normal genotype}
       \item{Alt}{The proportion of cells with only the CNA if the context C=C1 or with only the SNV if the context C=C2}
       \item{Both}{The proportion of cells with both the SNV and the SCNA}
       \item{Context}{Context at the mutation. If C1 then the SNV occurred after the SCNA, if C=c2 then the SNV occurred before the SCNA}
       \item{residual}{Residual after limSolve approximation.}
     }
     }
}
\description{
This function computes the cellular prevalence of a list of somatic mutations in
 cancer.  The function applies the model to a range of mutations located at a given genomic region or at the whole genome scale. 
The function invokes \code{\link{getPrevalence}}  to compute the cellular prevalence for each mutation of the list.
The model computes the prevalence of a somatic
  mutation relatively to close and eventually phased germline SNP as specified in \code{\link{getPrevalence}}.
}
\examples{

#Example 1: 

input_file=system.file("extdata","phylogeny1_d300_n80.tsv", package = "OncoPhase")
input_df<-read.table(input_file,header=TRUE)
rownames(input_df) = input_df$mutation_id
print(input_df)
#  mut_id varcounts_snv refcounts_snv major_cn minor_cn varcounts_snp refcounts_snp
#a      a      151  152        1        1      151  135
#b      b      123  176        1        1      161  150
#c      c       94  209        2        1      176  134
#d      d       23  283        1        1      155  144
#e      e       60  228        2        0      174  125

prevalence_df=getSamplePrevalence(input_df,nbFirstColumns = 1)

print(prevalence_df)
#  mut_id   Prev   Germ    Alt   Both Residual Context
#  a      a 0.9967 0.0017 0.0017 0.9967  3.1e-03      C1
#  b      b 0.8230 0.0890 0.0890 0.8230  1.3e-03      C1
#  c      c 0.4010 0.6000 0.0910 0.3100  3.9e-33      C2
#  d      d 0.1500 0.4200 0.4200 0.1500  1.4e-03      C1
#  e      e 0.2490 0.7500 0.0890 0.1600  5.1e-31      C2


#'@seealso \\code{\\link{getPrevalence}}
}

