% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoPhase_methods.R
\name{getSinglePrevalence}
\alias{getSinglePrevalence}
\title{Compute the cellular prevalence of each group of cells}
\usage{
getSinglePrevalence(varcounts_snv, refcounts_snv, major_cn, minor_cn,
  varcounts_snp, refcounts_snp, context, Trace = FALSE,
  LocusCoverage = TRUE, NormalCellContamination = NULL)
}
\arguments{
\item{varcounts_snv}{A count of 
alleles supporting the variant sequence of  the somatic mutation}

\item{refcounts_snv}{A count of 
alleles supporting the reference sequence  of  the somatic mutation}

\item{major_cn}{major copy number at the locus of the mutation}

\item{minor_cn}{minor copy number (or a vector of copy number if multiple tumor samples)}

\item{varcounts_snp}{A count of 
alleles supporting the variant sequence of  the Germline SNP}

\item{refcounts_snp}{A count of 
alleles  supporting the reference sequence  of  the Germline SNP}

\item{context}{represents either the situation of a mutation which occurred after the CNV ("C1") or the context of a mutation which occurred before the CNV ("C2"). If not provided, the right context will be estimated from the input}

\item{Trace}{Print a trace of the eecution.}

\item{LocusCoverage}{when set to TRUE, the SNV locus coverage is estimated to the average coverage of the phased SNP and the variant allele fraction is the ratio of the variant allele count over the estimated locus coverage. Default = TRUE.}

\item{NormalCellContamination}{If provided, represents the rate of normal cells contaminations in the experiment.}
}
\value{
A list of the three cellular prevalence of each of the three groups of cells
}
\description{
This is a generic function to compute the detailed prevalence of a single mutation using the linear system of  the model.
}
\examples{

Prevalences = getSinglePrevalence(3, 10,2,1,8,5,"C1")

 print(Prevalences)
# Germ  Alt Both 
# 0.4  0.0  0.6 

}
\seealso{
\code{\link{getPrevalence}},   \code{\link{getMatrices}}
}

