\name{OncoPhase-package}
\alias{OncoPhase-package}
\alias{OncoPhase}
\docType{package}
\title{
\packageTitle{OncoPhase}
}
\description{
\packageDescription{OncoPhase}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{OncoPhase}
\packageIndices{OncoPhase}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{OncoPhase}

Maintainer: \packageMaintainer{OncoPhase}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
