% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoPhase.R
\docType{package}
\name{OncoPhase}
\alias{OncoPhase}
\alias{OncoPhase-package}
\title{OncoPhase package for somatic mutations cellular prevalence estimation using haplotype phasing}
\description{
The main function for somatic mutation cellular prevalence computation is  \code{\link{getPrevalence}}. 
See the manual and examples at \code{\link{getPrevalence}} for more details.
}
\details{
Input data for simple case studies can be generated  with the function \code{\link{build_casestudy}}.

The package include experimental data for chromosome 10, 15, 18  and 22 for two patients retrieved from a parallel study.
 (see for example  \code{\link{chr15_OP1019}} and \code{\link{chr22_11152}} )

For more detailed information on usage, see the package vignette, by typing
\code{vignette("OncoPhase")}. All support questions should be emailed to the authors.
}
\author{
Donatien Chedom-Fotso, Ahmed Ahmed, Christopher Yau.
}
\references{
OncoPhase reference:

OncoPhase: A package for computing Somatic Mutation cellular Prevalence in cancer using haplotype phasing. Bioinformatics 2016. Submitted

OncoPhase reference:

” Ovarian cancer haplotype sequencing reveals ubiquitous SOX2 overexpression in the premalignant fallopian tube epithelium”
}
\keyword{package}

