% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\docType{class}
\name{Phylogeny-class}
\alias{Phylogeny}
\alias{Phylogeny-class}
\title{The Phylogeny class}
\description{
An S4 base class representing a phylogeny
}
\section{Slots}{

\describe{
\item{\code{Name}}{Object of class \code{\link{character}} representing the name of the phylogeny}

\item{\code{NbSNV}}{Object of class integer, attribute of the class Phylogeny representing the number of single nucleotide variations (SNVs).}

\item{\code{NbSCNAs}}{Object of class integer, attribute of the class Phylogeny representing the number of somatic copy number alterations (SCNAs)}

\item{\code{snv_ids}}{Object of class list, attribute of the class Phylogeny representing the identifiers or names of the NbSNVs SNVs clusters.}

\item{\code{scna_list}}{Object of class list, attribute of the class Phylogeny representing the list of SCNA given in the form list(scna_1_name=scna_1_attibutes, scna_2_name=scna_2_attibutes,...,scna_NbSCNAs_name=scna_NbSCNAs_attibutes ). 
Each SCNA attribute is a list containing 2 fields :
\describe{
\item{CN}{A pair of integer representing the major and minor copies numbers of the SCNA given in the form c(major, minor)}
\item{LOC}{ The location or genomic region affected by the SCNA represented in term of the list of SNVs spanned by the SCNA. LOC is given inform of a vector (a_1, a_2, ..., a_NbSNV\]). Each  a_i takes values in {0,1,2,3,4} as follow:
\describe{
\item{a_i=0}{the scna do not span the locus of the SNV i}
\item{a_i=1}{the SCNA span the SNV i locus, and the SNV is harbored by all the copies of the major copy number chromosome}
\item{a_i=2}{the SCNA span the SNV i locus, and the SNV is harbored by all the copies of the minor copy number chromosome}
\item{a_i=3}{ the SCNA span the SNV i locus, and the SNV is harbored by one copy of the major copy number chromosome}
\item{a_i=4}{the SCNA span the SNV i locus, and the SNV is harbored by one copy of the minor copy number chromosome}
}
}
}}

\item{\code{Clones}}{Object of class list, attribute of the Class Phylogeny containing the list of the clones given in the form list(clone_1_name=clone_1_attibutes, clone_2_name=clone_2_attibutes,...,clone_NbClones_name=clone_NbClones_attibutes). The Germline clones do not need to be list, it will be deduced. Each clone's attribute is a list containing the following fields :
\describe{
\item{snv}{ID list of the SNVs harbored by cells of the clone}
\item{scna}{ID list of the SCNAs affecting the  cells of the clone}
\item{prev}{Cellular prevalence of the clone}
}}
}}
\seealso{
\code{\link{simulation.Phylogeny}},
}

