% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoPhase.R
\name{getPrevalence}
\alias{getPrevalence}
\title{Somatic mutation cellular prevalence using haplotype phasing.}
\usage{
getPrevalence(snp_allelecount_df, ref_allelecount_df, phasing_association_df,
  major_copynumber_df, minor_copynumber_df, normalfraction_df,
  nbFirstColumns = 3, method = "oxford1", tumoursamples = NULL,
  region = NULL)
}
\arguments{
\item{snp_allelecount_df}{A data frame containing for each mutation the  allelic 
count of the variant at each tumour. Should contains at least Chrom (The mutation
 chromosome) , End (The mutation position) and IsGermline (is the mutation a Germline
  or Somatic)  among the first columnns.}

\item{ref_allelecount_df}{A data frame containing for each mutation the allelic count
of the reference at each tumour.Should contains at least Chrom (The mutation
 chromosome) , End (The mutation position) and IsGermline (is the mutation a Germline
  or Somatic)  among the first columnns.}

\item{phasing_association_df}{A data frame containing for each somatic mutation, 
a colon separated list of germline SNP phased to it.}

\item{major_copynumber_df}{A data frame containing for each mutation the major
chromosonal copy number at each tumour sample.}

\item{minor_copynumber_df}{A data frame containing for each mutation the minor
chromosonal copy number at each tumour sample.}

\item{normalfraction_df}{A data frame containing for each mutation the fraction of
normal cell contamination at each tumour/sample of the study.}

\item{nbFirstColumns}{Number of first columns in snp_allelecount_df to reproduce in
the output dataframe e.g: Chrom, Pos, Vartype.}

\item{method}{The method to be used (default : oxford1 , alternatives are oxford2 
, oxford3.)}

\item{tumoursamples}{The list of tumour samples to consider for the prevalence
computation. This samples should be present as column header in the data frames 
 snp_allelecount_df, ref_allelecount_df, major_copynumber_df,minor_copynumber_df
 and  normalfraction_df. If not provided, the headers from nbFirstColumns + 1 to 
 the last column of snp_allelecount_df is retrieved and its intersection with the
  headers of the four othe rmatrices is considered as the tumour samples
 of the allelecount or copy number matrices)}

\item{Region}{The region of the genome to consider in the format chrom:start-end 
e.g "chr22:179800-98767}
}
\value{
A data frame containing the cellular prevalence of each mutation at each tumour sample.
}
\description{
This is a generic function computing the cellular prevalence of somatic mutations in
 cancer using haplotype phasing. The model compute the prevalence of a somatic
  mutation relatively to close and phased germline mutation. It uses three sources
   of information as input : The allelic counts, the phasing information and the 
   copy number alteration.
}
\examples{

# Example 1Loading an example data of two somatic mutations, 5 germlines SNP, and 3 tumour samples

data(simpleExample2)
attach(simpleExample2)
prevalence_df=getPrevalence(snp_allelecount_df, ref_allelecount_df, phasing_association_df, major_copynumber_df,minor_copynumber_df,normalfraction_df,nbFirstColumns,tumoursamples)

#Example 2 : Computing somatic mutation cellular prevalence on chromosome 15 of patient 11152 (data retrieved from a parallele study)

data("chr15_11152")
attach(chr15_11152)
masterprevalence_df=getPrevalence(snp_allelecount_df, ref_allelecount_df, phasing_association_df, major_copynumber_df,minor_copynumber_df,normalfraction_df,nbFirstColumns=6,region="chr15:50000000-80000000")

# Example 3 : Creating a simple example with one somatic mutation and one germline mutation on a single tumour sample

#Empty dataframe
snpcount_df=as.data.frame(matrix(ncol=4,nrow=2))
names(snpcount_df) = c("Chrom","End","IsGermline","Tumour1")
rownames(snpcount_df) = c("mutation1","mutation2")
refcount_df = snpcount_df
major_cn_df= as.data.frame(matrix(ncol=1,nrow=2))
names(major_cn_df) = "Tumour1"
rownames(major_cn_df) = c("mutation1","mutation2")
minor_cn_df = major_cn_df
normalfraction_df = major_cn_df

#Filling the dataframes
snpcount_df["mutation1",] = c("chr1", 200100,0,40)
snpcount_df["mutation2",] = c("chr1",  200900,1,60)
refcount_df["mutation1",] = c("chr1",  200100,0,20)
refcount_df["mutation2",] = c("chr1",  200900,1,40)
major_cn_df["Tumour1"] = c(1,1)
minor_cn_df["Tumour1"] = c(1,1)
normalfraction_df["Tumour1"] = c(0.2,0.2)

#Phasing association
phasing_association_df = as.data.frame(matrix(ncol=1,nrow=1))
colnames(phasing_association_df) = c("PhasedGermline")
rownames(phasing_association_df) = c("mutation1")
phasing_association_df["mutation1","PhasedGermlines"] = "mutation2"

#Computing the prevalence
prevalence_df=getPrevalence(snpcount_df, refcount_df, phasing_association_df, major_cn_df,minor_cn_df,normalfraction_df)

print(prevalence_df)

#          Chrom    End IsGermline   Tumour1
# mutation1  chr1 200100          0 0.6666667


}

