% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OncoPhase.R
\docType{package}
\name{OncoPhase-package}
\alias{OncoPhase-package}
\title{OncoPhase package for somatic mutation cellular prevalence estimation using haplotype phasing}
\description{
The main functions for somatic mutation cellular prevalence computation is  \code{\link{getPrevalence}}. See the examples at \code{\link{getPrevalence}} for basic steps.
For more detailed information on usage, see the package vignette, by typing
\code{vignette("OncoPhase")}. All support questions should be emailed to  the authors.
}
\author{
Donatien Chedom-Fotso, Ahmed Ahmed, Christopher Yau.
}
\references{
OncoPhase reference:

OncoPhase: A package for computing Somatic Mutation cellular Prevalence in cancer using haplotype phasing. Bioinformatics 2016. Submitted

OncoPhase reference:

SOX2 paper. Detail of the reference to be added
}
\keyword{package}

